Add product variant for admin

+++ app/Http/Controllers/Admin/ProductController.php
@@ -8,6 +8,7 @@ use App\Models\Brand;
 use App\Models\Category;
 use App\Models\ChildCategory;
 use App\Models\Product;
+use App\Models\ProductVariant;
 use App\Models\SubCategory;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
@@ -191,6 +192,12 @@ class ProductController extends Controller
             abort(404);
         }
         $product = Product::findOrFail($id);
+        $variants = ProductVariant::where('product_id', $product->id)->count();
+        if($variants > 0){
+            notyf()->error('There are variants under this product. Please delete them first!');
+            return response(['status' => 'error']);
+        }
+
         if (File::exists(public_path($product->thumb_image))) {
             File::delete(public_path($product->thumb_image));
         }
